academic_literature_review_RAG/requirements.txt
langchain==0.0.184
openai==0.27.8
chromadb==0.3.26
tiktoken==0.4.0
gradio==3.35.2
PyPDF2==3.0.1
pyyaml==6.0
